browser-ext: refactor Background component
